requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
deta==1.0.0
fastapi==0.68.0
h11==0.12.0
httptools==0.2.0
idna==3.2
pydantic==1.8.2
python-decouple==3.4
PyYAML==5.4.1
requests==2.26.0
soupsieve==2.2.1
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.15.3
watchgod==0.7
websockets==9.1
